requirements.txt
streamlit
requests
pydantic
python-dotenv
faiss-cpu
groq
